+++ src/composables/getCVDModeFromURL.js
@@ -1,0 +1,8 @@
+import getURLParam from "@/composables/getURLParam";
+
+const VALID_MODES = ['normal', 'protanopia', 'deuteranopia', 'tritanopia'];
+
+export default function getCVDModeFromURL() {
+  const mode = getURLParam("cvdMode");
+  return VALID_MODES.includes(mode) ? mode : false;
+}